fixDirectories.
So nutshell version: in the library, you're expecting to see stuff like

  LIBRARY/foo/bar.png
  LIBRARY/foo/baz.png

All fine and dandy. Hooray. And it refers to those within the xml with
slashes as directory seperators as God intended.

But it appears that on WINDOWS, where the file seperator is a backslash,
it winds up writing out:

  LIBRARY/foo\bar.png
  LIBRARY/foo\baz.png

That's right; they're in the LIBRARY directory, with filenames containing
a backslash. At which point, it appears that a reference to 'foo/bar.png' in
the XML gets turned into a file open request for 'foo\bar.png' since you're
on windows, and all works as expected. "Yay"

Except, of course, now you've got these weird not-really-in-a-directory files.
And you share your fla with someone on a mac. and so it looks for 'foo/bar.png'
and DOESN'T get your new one. And you're shadowing symbols and everything goes
crazy and people argue over what's really in this file.

God DAMNIT, Adobe.